fix(types): add RawSlots in h signature
`RawSlots` is supported as a third argument in most signatures except when the first one is a `string`. I _think_ this is a valid use-case, as @dobromir-hristov used it for stubs in VTU next.
We are currently forced to cast the parameter as `any` to make TS happy in VTU. This commit should fix that.