fix(runtime-core): vnode.el is null in watcher after rerendering
An instance’s `el` property can be `null` inside a watcher after a component was rerendered (e.g. as a result of a changed prop). In an instance’s `update` method, the branch for handling mounted instances doesn’t set the `el` property of what will become the instance’s `vnode` property soon enough. Upon processing of the `updateComponentPreRender` function, the `el` property is then `null` which causes issues in watcher routines that rely on an instance’s `$el` property.