feat(computed): add cleanup method
The cleanup method will remove all dep references to and from dependency sets. This allows the computed to be GC collected (if no other references remain).

Another reason to call this cleanup is to reduce the number of dependencies. If there are a large number of dependencies, triggering a reactive object may become unnecessarily as all deps need to be triggered.