fix(runtime-core): InjectionKey compatible with symbol type
The current type of `InjectionKey` does not allow to use it as an ES computed property.
This is problematic as the API of `mount` in VTU lets you add providers with `provide: { key: value }`.
If you want to use a key defined as:

    export const storeKey: InjectionKey<Store<State>> = Symbol();

Then you currently have to cast the key as symbol:

    const wrapper = mount(App, { global: { provide: { [storeKey as symbol]: store } } });

This commit allows to use:

    const wrapper = mount(AppWithVuex, { global: { provide: { [storeKey]: store } } });