fix(compiler): Addressed further line-ending issues
There were additional places that we needed to take line ending sequences into account to ensure that we placed  the ^^^s exactly where they should be.
Unix and Windows should now produce identical results.

Additionally, code was modified to handle source that is passed in that does not have a trailing newline character before the end of the source content.