fix(runtime-dom): handle Symbol attributes
Currently, passing a `Symbol` to an attribute throws in `runtime-dom` with:

```
TypeError: Failed to execute 'setAttribute' on 'Element': parameter 2 is a symbol, which cannot be converted to a string.
```

This is because the underlying library used for conversion does not handle Symbol (see https://github.com/jsdom/webidl-conversions/issues/14).

This issue cascades into VTU-next as `shallowMount` attempts to write props values as attributes.
See https://github.com/vuejs/vue-test-utils-next/issues/1076 for more context.