chore(@vue/shared): add `"sideEffects": false`
As far as I can tell, all the code in this package is side-effects-free.

Without the `sideEffects` flag, bundlers rely on code analysis and
the `/*#__PURE__*/` comments for tree-shaking.
As bundlers' implementations vary, the analysic may or may not be
exhaustive.

For example, a file that contains only the following line of code:
```js
import { capitalize } from '@vue/shared'
```
If bundled by esbuild (`esbuild --bundle test.js --tree-shaking=true`),
the output would be tens of lines.

After this PR, the output will be an empty IIFE.

Credit to @hardfist for discovering this issue.