Add custom named slots into customElements/WebComponents
I added named slots. We needed to add slots to the web-component as it does in vue. After the changes have been made, we can add a template with the name and content. A slot inside the component will be rendered.

Example of use

index.html
<ce-epicrisis id = "testWidget">
    <template ce-right-slot-header>
        <span style = "color: # af72dd;"> __ Example header slot__ </span>
    </template>
    <template ce-footer-slot>
        <span style = "color: # af72dd;"> __ Example header slot__ </span>
    </template>
</ce-epicrisis>

component.vue
  <div class = "header-wrapper" data-provide-ce-right-slot-header>
    <h2 class = "widget-title"> {{translateTitle}} </h2>
    <slot name = "ce-right-slot-header"> </slot>
  </div>

In the web-component, add 'template' with the attribute by which it will be recognized - e.g. ce-right-slot-header

In the component, add an attribute starting with 'data-provide' plus the name we provided in the template, that is: data-provide-ce-right-slot-header

On this basis, the web-component will be rendered and the value from template will be embedded in it.