fix(jsx-runtime): fix automatic runtime implementation
The JSX automatic runtime consists of two modules. These modules require
different exports.

- `vue/jsx-runtime` is for production. This has the following exports:
  - `jsx`
  - `jsxs`
  - `Fragment`
- `vue/jsx-dev-runtime` is for development. This has the following exports:
  - `jsxDEV`
  - `Fragment`

Whereas a JSX pragma of the classic runtime accepts children as the 3rd..nth
arguments, in the automatic runtime this is passed as the `children`
prop. If the JSX element has one child, the `jsx` function is used and
`children` is that child. If the JSX element has more, the `jsxs`
function is called and `children` is an array of all childrne. If no
children are passed, this prop is omitted.

The third argument is the key. In the classic runtime this was part of
props.

If the development transform is used, `jsxDEV` from the
`vue/jsx-dev-runtime` module is used intead. This is similar to `jsx` /
`jsxs`, but it accepts more arguments.

The 4th argument determines if the production runtime would use `jsxs` or
`jsx`.

The 5th argument is positional information that you could use to enhance the
developer experience. Note that different compilers compile the file
name differently. (relative vs absolute paths)

The 6th argument is the value of `this` in the scopewhere the JSX element was
created. React uses this to warn about string refs.

This implementation combines the production and development runtime in
the same file, and uses an package exports to expose both import entry
points.