fix: correctly resolve types from relative paths on Windows
Closes #8671
Closes https://github.com/vuejs/vue-loader/issues/2048

`dirname` is the safest way to get a directory name from a path string.

As for this specific bug, it's because
`path.posix.join(AnyWindowsPath, '..')` returns `.`.

I'm not sure whether there will be unexpected regression if I modify the
`joinPaths` utility, so I changed the smallest possible bit of code to
fix this.