fix(server-renderer): Fix renderComponentVNode to call serverPrefetch option w/async setup
This fixes a bug where the serverPrefetch option would not be called
with an async setup method.

A `prefetches` reference was created while `instance.sp` was `null`.
However, with an async setup, `instance.sp` has the serverPrefetch
methods only after setup resolves.

The fix is achieved by waiting to for setup to resolve before trying to
access the the serverPrefetch options from `instance.sp`.