fix(transition-group): run `forceReflow` on the correct document (fix #13849)
Previously, `forceReflow()` would always use `document`.
When a component with a `TransitionGroup` was moved to another document, move transitions broke because of this.
This is fixed by accessing the current instance's `ownerDocument` and using that for the reflow.
This is my first time contributing to vue, and I'm not entirely sure if this fix is the right way to go (especially when it comes to perf and possibly unintended side effects).